backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
PyJWT
passlib[bcrypt]
python-multipart
